Add EmailSenderTest
