--- src/test/java/com/ericsson/ei/notifications/MailSenderTest.java
@@ -1,100 +1,0 @@
-/*
-   Copyright 2019 Ericsson AB.
-   For a full list of individual contributors, please see the commit history.
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-       http://www.apache.org/licenses/LICENSE-2.0
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-package com.ericsson.ei.notifications;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.Mockito.*;
-import static org.powermock.reflect.Whitebox.invokeMethod;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.HashSet;
-import java.util.Set;
-
-import javax.annotation.PostConstruct;
-import javax.mail.internet.MimeMessage;
-
-import org.apache.commons.io.FileUtils;
-import org.json.JSONObject;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.mail.MailException;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.MvcResult;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-
-import com.ericsson.ei.App;
-import com.ericsson.ei.controller.model.QueryResponse;
-import com.ericsson.ei.jmespath.JmesPathInterface;
-import com.ericsson.ei.notifications.EmailSender;
-import com.ericsson.ei.notifications.HttpRequestSender;
-import com.ericsson.ei.notifications.InformSubscriber;
-import com.ericsson.ei.subscription.RunSubscription;
-import com.ericsson.ei.handlers.MongoDBHandler;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.mongodb.MongoClient;
-
-import de.flapdoodle.embed.mongo.distribution.Version;
-import de.flapdoodle.embed.mongo.tests.MongodForTestsFactory;
-import com.ericsson.ei.exception.NotificationFailureException;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringBootTest(classes = { App.class })
-@AutoConfigureMockMvc
-public class MailSenderTest {
-
-    @Autowired
-    private EmailSender emailSender;
-
-    @Mock
-    private JavaMailSender javaMailSender;
-
-    @Mock
-    MimeMessage mimeMessage;
-
-    @Before
-    public void beforeTests() {
-    }
-
-    @Test(expected=NotificationFailureException.class)
-    public void sendEmailThrowsException() throws Exception {
-        emailSender.sendEmail(mimeMessage);
-    }
-
-    @Test
-    public void prepareEmailTest() throws Exception {
-        emailSender.prepareEmailMessage("My test message", "nisse@manpower.com", "My Subject");
-    }
-
-}
